Base project upload
Upload the AgeOfChampions code as a base of this project, duplicated, a
simple and advanced folder. The simple folder is where most of the code
will be deleted/simplified and the advanced folder is a backup of the
code in case I want to do a more indepth guide on task managment.